backend/channels/pyproject.toml
[tool.poetry]
name = "channels"
version = "0.1.0"
description = "Channels"
authors = ["DmitriyKhudiakov"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.94.0"}
pytest = "^7.2.2"
pytest-asyncio = "^0.20.3"
motor = {extras = ["all"], version = "^3.1.1"}
redis = "^4.5.1"
pika = "^1.3.1"
aio-pika = "^9.0.5"
flower = "^1.2.0"
celery = "^5.2.7"
passlib = "^1.7.4"
jinja2 = "^3.1.2"
python-jose = "^3.3.0"

[tool.pytest.ini_options]
pythonpath = [
    ".", "src"
]
filterwarnings = ["ignore:SelectableGroups dict interface is deprecated. Use select.:DeprecationWarning"]
asyncio_mode="auto"
